backend/requirements.txt
Django==5.0.0
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5
qrcode[pil]==7.4.2
python-decouple==3.8
Pillow==10.1.0
